Add install job for single collection download
